tests and ruff passes
